refactor: reorganize user-related hooks into a dedicated folder
- Move user hooks (`useManagement`, `useGeographicData`, `useEditInitialization`, `useAddressData`, `useUserDetailsForm`) to `hooks` directory for better structure
- Update related imports across components for consistency and maintainability
- Remove unnecessary blank lines in `validation.ts` file to improve readability
